// apps/web/src/shared/api/generated/model/deliveryReadControllerHistoryListTab.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type DeliveryReadControllerHistoryListTab =
  (typeof DeliveryReadControllerHistoryListTab)[keyof typeof DeliveryReadControllerHistoryListTab];

export const DeliveryReadControllerHistoryListTab = {
  all: "all",
  active: "active",
  received: "received",
  cancelled: "cancelled",
} as const;